Add support for post-logging
By default ecAudit use pre-logging by logging an ATTEMPT before passing
the operation down to the QueryProcessor. In the event of a failure, a FAILURE
message is also logged.

With post-logging only one message (SUCCEEDED or FAILED) will be logged after
the operation has completed. This is similar to the C* 4.0 approach.

Closes #24