Configurable chronicle fields
New chronicle configuration parameter - "fields" make it possible to select
which audit fields that should we written to the binary file.

The audit record version is stepped and a fields parameter is added (containing
a bitmask indicating which fields are present in the record).

When ecLog reads/prints the log, all available fields will be printed.

Closes #101